Move to logger package
