+++ logger/src/main/scala/echopraxia/plusscala/logger/LoggerFactory.scala
@@ -1,4 +1,4 @@
-package echopraxia.plusscala
+package echopraxia.plusscala.logger
 
 import echopraxia.logging.spi.{Caller, CoreLoggerFactory}
 import echopraxia.plusscala.api.FieldBuilder
