+++ logger/src/test/scala/echopraxia/plusscala/logger/ConditionSpec.scala
@@ -1,9 +1,9 @@
-package echopraxia.plusscala
+package echopraxia.plusscala.logger
 
 import ch.qos.logback.classic.LoggerContext
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.core.read.ListAppender
-import api._
+import echopraxia.plusscala.api.FieldBuilder
 import echopraxia.plusscala.logging.api.{Condition, JsonPathCondition, Level}
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.funspec.AnyFunSpec
